use cost estimation
this allows us to minimize the number of ids that need to get filtered
by the filters, while at the same time still taking the execution cost
of each filter into account.
f.e. if we have the choice between first calling a filter which uses
regex or first calling a filter that doesn't use regex, we'd first want
to call the one that's not using regex and let it reduce the size of the
potential result set, that way the regex-using filter would later only
get applied on a smaller set of potential results.