avoid race condition when returning interned values from index
this copies archives before returning them from the index methods
AddOrUpdate or Update and increases the refcounts of all their
interned members. before they go out of scope the refcounts get
decreased accordingly.
that way we can prevent that something gets deleted from the object
store while we still hold a reference to it.
to make the additional copying cheaper it uses pools of resources.