better verification on meta tag record upsert
when receiving a new meta tag record upsert request we shouldn't only
validate the given query expressions, but also that the given set of
expressions can be used to instantiate a query. if that's not the case,
then the given meta tag record should be considered invalid and should
get rejected.