when adding new point to aggregate, always see if we can flush
e.g. when flushing every 10s and we reach a timestamp like 12:34:10
we can always flush, because any next point will always go into
the next bucket.  We already did this but only in the case of already
having data in the bucket. In the case of an unfortunate mismatch
between aggergation and series (e.g. aggregate across 10s but the
data itself comes every 10s or even less frequent), we can also
just flush in this case, rather than waiting.