process requests on a per shard basis, not per peer.
This change refactors the processing flow for render queries.
Matching series are fetched from all shards, but instead of tracking
the node that the series were found on, we now track the shard.
When processing requests, we group by the shard, then send all of
those requests to a peer from the shard. if the request to the peer
fails, additional peers for the shard will be tried.