remove from argument from auxiliary tag api calls
the from argument has never been documented and is not used by graphite,
so it should be safe to assume that nobody is using it. keeping it would
make the queries quite a lot more complicated once we have to also take
meta tags into account, so we remove it.