only start filters when needed
since we still always need to filter by the from timestamp, we are now
doing this in the id selector by calling a new method on the tag query
context called newerThanFrom.
to make this change work it was necessary to pass the result channel
into getIds(), instead of returning it from it. this resulted in a bit
of rewiring of the channels and especially where channels get closed.
to make it possible to close the result channel when the id selector is
finished, getIds() is now a blocking method.