rewrite enricher to use lookup map for enrichment
this replaces the enricher implementation to improve its enrichment
performance. it drops the enrichment cache and it also doesn't filter
metrics based on the meta records at the enrichment stage anymore.
instead it is now building a map from which it can lookup the metric
keys and resolve them into meta records, from which it then gets the
meta tags.
especially in scenarios where there is a large number of meta records in
the index this performs much better than the old implementation.