Loop new request identifiers/differentiators through request path
if multiple queries are given
(e.g. target=foo*&target=sum(foo*)&target=summarize(foo*))
we need to be able to differentiate them. We wouldn't want to conflate
MDP or PN-optimized data when we need as high resolution data as
possible, for example.

We already differentiate expr.Req, models.Req and models.Series
based on query, from, to, consolidator, but now we need to do it by
the requests' MDP and PNGroup as well.

Note that FuncGet() is setup with an expr.Req early on in the
request parsing phase (NewPlan()), whereas the data it will load
later on first gets turned into a models.Req in executePlan() and
then in models.Series in getTargets(). Then from those series
we parse out again what the original expr.Req was and classify
the data by expr.Req when we issue plan.Run(), such that FuncGet()
is able to load the data based on the expr.Req it has.

Note also that models.Req and models.Series may be transmitted
over the network.

We als update mergeSeries accordingly to not conflate optimized
and non-optimized data for the same metric together.
(see query example above).

models.Req :
 - add PNGroup
 - be more explicit that we take values from expr.Req
 - we can't import expr, so:
   * refactor how we create it (we can't take in expr.Req as an arg
   * move PNGroup from expr to models

models.Series:
  - add MDP and PNGroup