summarize(): in !alignToFrom case, return output in canonical format
See https://github.com/grafana/metrictank/issues/1811 for more details
Before, it was common for output to contain an extraneous point that
lies before 'from' (and wouldn't be rendered by Grafana in graph
panels), which made the output invalid.

This would result in trouble when

* combining such a series with another series
  (https://github.com/grafana/metrictank/issues/1811)
* or when needing normalization
  (https://github.com/grafana/metrictank/issues/1845)

This effectively slightly changes the output format, but is more robust.