support fetchFunc having awareness of cluster peers
fetchFunc can use this to determine the ratio of how much
data the target peer owns compared to the cluster as a whole.
Caveat: this is based on live cluster state. If shardgroups
go completely down it'll look like the target peer owns more of the
cluster than it actually does.
if query limits are set based on this, the limits would loosen up as
shards leave the cluster.