bugfix: don't apply limit before from-filtering, make find iterator
OLD:
* find() does a complete find (without 'from' filtering)
  these results can nicely be cached.
* In Find() we do 'from' filtering
in the last few PR's we implemented the "series limiter" in find
because we don't want to first assemble the entire result set only to
then check the limit, as this makes the limit moot, effectively.
but implementing the limiter prior to from-filtering makes it
impossible to apply correctly, as typically most entries will not be
included in the final resultset due to the 'from' filtering.

NEW:

The new approach is a lockstep approach with an iterator.

We want to avoid loading too much data in RAM (apply limit as early
possible), while caching the find response body in a way that's
not affected by the 'from' filter.

We do this by changing find to an iterator, which feeds data to its
caller (Find) which can apply the limit while 'find' is running.
When the iteration is completed, the iterator can still add the full
find output to the cache.

This also means that find iteration is now depth first rather than
breadth first. we compute all matchers up front which is a bit more
upfront cost but the benefit to bail out early, downside: may compute
matchers needlessly if we would otherwise have found out "early" that a
query had no matches

To make find logic depth first without changing too much code
- since we cannot "yield" from inside the find function -
we need to use channels to feed results to the iterator. Not the most
beautiful approach...