Detailed plan errors
Right now when we fail the plan we see an error message which is:

    argument bad type. expected int - got etString

If the query is not trivial (with nested function calls, for instancer),
it becomes very complicated to understand which arg it's referring too.

This adds context to the errors returned by NewPlan() (specifically,
newplanFunc()), telling which function and which arg caused the error.

The original error is wrapped now so all the usages of NewPlan() now
check the wrapped error rather than specific type.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>